[feat] fix bottom nav

+++ app/src/main/java/com/klimaatmobiel/domain/Group.kt
@@ -5,7 +5,7 @@ import kotlinx.android.parcel.Parcelize
 
 @Parcelize
 
-class Group(val groupId: Long, val groupName: String, val projectId: Double, val project : Project,
+class Group(val groupId: Long, val groupName: String, val projectId: Long, val project : Project,
             var order : Order, val uniqueGroupCode: String, val pupils: MutableList<Pupil>) : Parcelable {
 
     fun findOrderItemById(orderitemId : Long) : OrderItem?{
